requirements.txt
PyGithub==1.58.0
requests==2.31.0
